change function name for readablity and add rust doc to the function
